+++ src/renderer/containers/SpectatePage/InfoBlock.tsx
