+++ src/renderer/containers/SpectatePage/SpectateItem.tsx
