+++ src/renderer/containers/SpectatePage/SpectatorIdBlock.tsx
